+++ backend/src/main/java/ee/lio/birthcalculator/dto/request/BenefitRequest.java
@@ -4,12 +4,13 @@ import com.fasterxml.jackson.annotation.JsonFormat;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public record BenefitRequest(
         @NotNull
         @Positive
-        Double salary,
+        BigDecimal salary,
 
         @NotNull
         @JsonFormat(pattern = "yyyy-MM-dd")
